add support (index, references) to wt collections (wt2g & wt10g)
