+++ docs/dumpindex-reference.md
@@ -8,6 +8,8 @@ Collection     | # of docs   | # of docs with content |  total terms | unique te
 Disk12         | 741676      | 741675                 |  219327909   | 950715       |
 AQUAINT        | 1031455     | 1031326                |  317703234   | 966882       |
 Disk45         | 528030      | 528030                 |  174540587   | 923435       |
+WT2G           | 245715      | 245679                 |  181774134   | 1653144      |
+WT10G          | 1688402     | 1688290                |  752326031   |              |
 Gov2           | 25172934    | 25170664               |  17343119816 |              |
 CW09B          | 50220189    | 50220159               |  31270685466 |              |
 CW12           |             |                        |              |              |
